requirements.txt
Flask==3.0.0
tensorflow==2.16.1
numpy==1.26.3
pillow==10.2.0